% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna_env.R
\name{load_all_rna_seq_envs}
\alias{load_all_rna_seq_envs}
\title{Load all RNA-Seq environments}
\usage{
load_all_rna_seq_envs(sex, base_results_path = "./results/rna_seq")
}
\description{
Load all RNA-Seq environments
}
