% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cage_env.R
\name{load_cage_peak_annotation}
\alias{load_cage_peak_annotation}
\title{Loading the annotation of peaks as a gRange object}
\usage{
load_cage_peak_annotation()
}
\description{
Loading the annotation of peaks as a gRange object
}
