% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna_env.R
\name{load_foo_rna_seq_env}
\alias{load_foo_rna_seq_env}
\title{Generic RNA-Seq Environment Loader}
\usage{
load_foo_rna_seq_env(sex, base_results_path = "./results/rna_seq/")
}
\arguments{
\item{sex}{Character string indicating sex ("female" or "male")}

\item{base_results_path}{Character string for base results directory}
}
\description{
Generic RNA-Seq Environment Loader
}
