% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant_env.R
\name{load_quant_seq_env}
\alias{load_quant_seq_env}
\title{Quant-Seq Environment}
\usage{
load_quant_seq_env()
}
\description{
Loads the Quant-Seq specific environment
- counts of peaks
- paths for:
 - results
 - deseq
}
