% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{process_overlapping}
\alias{process_overlapping}
\title{Generic function to find overlaps between two sets of genomic ranges
and summarize the results.}
\usage{
process_overlapping(
  query_ranges,
  subject_ranges,
  id_col = NULL,
  extend_bp = 0,
  col_query_to_subj = NULL,
  verbose = TRUE
)
}
\arguments{
\item{query_ranges}{GRanges object representing the query features (e.g., gene TSSs)}

\item{subject_ranges}{GRanges object representing the subject features (e.g., TEs)}

\item{id_col}{Optional string specifying the column name in mcols(query_ranges) that uniquely identifies the entity (e.g., gene ID). If NULL, all distinct query features are considered.}

\item{extend_bp}{Integer specifying the number of base pairs to extend the query ranges on both sides (default is 0, meaning no extension).}

\item{verbose}{Logical indicating whether to print summary statistics (default is TRUE).}
}
\value{
A list containing:
  - query_with_hit: GRanges of query features that have overlaps
  - subject_hits: GRanges of subject features that overlap with the query
  - stats: A list with total number of query features, number of unique entities with hits, and percentage with hits.
}
\description{
Generic function to find overlaps between two sets of genomic ranges
and summarize the results.
}
\note{
# This function requires the GenomicRanges package.

subject_ranges <- GRanges(seqnames = rep("chr1", 3),
                          ranges = IRanges(start = c(1,11,21), end = c(10,20, 30)),
                          strand = c("+", "+", "+"),
                          gene_id = paste0("gene", 1:3),
                          score = 1:3)
query_ranges <- GRanges(seqnames = rep("chr1", 3),
                        ranges = IRanges(1:3, width=1),
                        strand = c("+", "+", "+"),
                        gene_id = paste0("gene", 1:3),
                        score = 1:3)

process_overlapping(query_ranges, subject_ranges) 
This returns that 100 % of the query features overlap with the subject features. Because all three TSS (
query_ranges) overlap with the subject ranges.

#' process_overlapping(subject_ranges, query_ranges)
This returns that 33.33 % of the query features overlap with the subject features. Because only one of the 
three subject ranges (subject_ranges) overlap with the query ranges.
}
\examples{
# Example usage:
results <- process_overlapping(query_ranges = gene_tss_gr, subject_ranges = teRanges, id_col = "ensembl_gene_id", extend_bp = 200)
}
